refactor(#160): relocate _text_width to _core (P1, prerequisite)
Move the cached rendered-width measurement _text_width out of make_drawing.py
into _core.py. The sheet/layout estimators call it 8×, so having it in _core
(below make_drawing) removes the would-be sheet -> make_drawing import cycle that
P3 (sheet.py) would otherwise hit. _core already depends on fonts/build123d, so
the move adds no new edge.

Pure move: make_drawing re-imports it (the _text_width unit test and the sheet
estimators keep working). Golden gate passes against the committed snapshots with
no regeneration; ruff check + format + mypy clean; text-width/estimator tests pass.
Adversarially checked: function body byte-identical, _core imports standalone (no
cycle), measurements unchanged, lru_cache intact.
